+++ src/NEUIK_HGroup.c
@@ -606,7 +606,7 @@ SDL_Texture * neuik_Element_Render__HGroup(
 		eNum = 6;
 		goto out;
 	}
-	xFree = rSize->w; /* free X-px: start with the full width and deduct as used */
+	xFree = (float)(rSize->w); /* free X-px: start with the full width and deduct as used */
 
 	/*------------------------------------------------------------------------*/
 	/* Check to see if the requested draw size of the element has changed     */
@@ -717,7 +717,7 @@ SDL_Texture * neuik_Element_Render__HGroup(
 		}
 
 		/* calculate the number of horizontal px per 1.0 of HScaling */
-		hFillPx = (int)((float)(xFree) / tScale);
+		hFillPx = (float)((int)((float)(xFree) / tScale));
 
 		/*--------------------------------------------------------------------*/
 		/* Render and place the child elements                                */
@@ -752,7 +752,7 @@ SDL_Texture * neuik_Element_Render__HGroup(
 				eNum = 4;
 				goto out;
 			}
-			xSize = rs.w;
+			xSize = (float)(rs.w);
 
 			/*----------------------------------------------------------------*/
 			/* Check for and apply if necessary Horizontal and Veritcal fill  */
@@ -766,7 +766,7 @@ SDL_Texture * neuik_Element_Render__HGroup(
 			{
 				/* This element is configured to fill space horizontally */
 				xSize = hFillPx * (eCfg->HScale) - (eCfg->PadLeft + eCfg->PadRight);
-				rs.w  = xSize;
+				rs.w  = (int)(xSize);
 			}
 
 			/*----------------------------------------------------------------*/
@@ -787,7 +787,7 @@ SDL_Texture * neuik_Element_Render__HGroup(
 					break;
 			}
 
-			rect.x = xPos + eCfg->PadLeft;
+			rect.x = (int)(xPos + eCfg->PadLeft);
 			rect.w = rs.w;
 			rect.h = rs.h;
 			rl.x = (eBase->eSt.rLoc).x + rect.x;
